+++ src/app/(tabs)/library/styles.ts
@@ -1,0 +1,5 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({});
+
+export default styles;
