+++ src/app/(tabs)/search/styles.ts
@@ -1,0 +1,48 @@
+import { Dimensions, StyleSheet } from 'react-native';
+
+const { width, height } = Dimensions.get('window');
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: 'white',
+    justifyContent: 'flex-start',
+    paddingLeft: 24,
+    paddingRight: 24,
+    paddingTop: 20,
+    gap: 14,
+  },
+  searchContainer: {
+    backgroundColor: 'transparent',
+    borderRadius: 10,
+    borderColor: 'transparent',
+    padding: 0,
+    marginBottom: 16,
+  },
+  inputContainer: {
+    backgroundColor: '#D9D9D9',
+    margin: 0,
+    borderRadius: 10,
+  },
+  greyOverlay: {
+    flex: 1,
+    position: 'absolute',
+    left: 0,
+    top: 0,
+    opacity: 0.2,
+    backgroundColor: 'black',
+    width,
+    height,
+    zIndex: 1,
+  },
+  noOverlay: {
+    flex: 1,
+    position: 'absolute',
+    left: 0,
+    top: 0,
+    width,
+    height,
+  },
+});
+
+export default styles;
