+++ src/app/(tabs)/story/styles.ts
@@ -1,0 +1,119 @@
+import { StyleSheet } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: 'white',
+    alignItems: 'flex-start',
+    justifyContent: 'flex-start',
+    paddingLeft: 24,
+    paddingRight: 24,
+    paddingTop: 48,
+  },
+  image: {
+    width: '100%',
+    height: 153,
+    marginBottom: 16,
+  },
+  authorImage: {
+    backgroundColor: '#D9D9D9',
+    width: 21,
+    height: 21,
+    borderRadius: 100 / 2,
+  },
+  title: {
+    fontFamily: 'Avenir',
+    fontSize: 24,
+    fontWeight: '400',
+    textAlign: 'left',
+    color: 'black',
+    marginBottom: 16,
+  },
+  author: {
+    display: 'flex',
+    flexDirection: 'row',
+    gap: 10,
+    marginBottom: 16,
+  },
+  authorText: {
+    fontFamily: 'Avenir',
+    fontSize: 12,
+    fontWeight: '400',
+    textAlign: 'left',
+    color: 'black',
+  },
+  genres: {
+    display: 'flex',
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    borderRadius: 10,
+    marginBottom: 16,
+  },
+  genresBorder: {
+    backgroundColor: '#D9D9D9',
+    paddingTop: 5,
+    paddingBottom: 5,
+    paddingLeft: 10,
+    paddingRight: 10,
+    borderRadius: 10,
+    marginRight: 8,
+  },
+  genresText: {
+    fontFamily: 'Avenir',
+    fontSize: 12,
+    fontWeight: '400',
+    color: 'black',
+    backgroundColor: '#D9D9D9',
+  },
+  shareButtonText: {
+    fontFamily: 'Avenir',
+    fontSize: 12,
+    fontWeight: '400',
+    textAlign: 'left',
+    color: 'black',
+    textDecorationLine: 'underline',
+    backgroundColor: '#D9D9D9',
+  },
+  excerpt: {
+    fontFamily: 'Avenir',
+    fontSize: 16,
+    fontWeight: '400',
+    textAlign: 'left',
+    color: 'black',
+    paddingTop: 16,
+    paddingBottom: 16,
+  },
+  story: {
+    fontFamily: 'Avenir',
+    fontSize: 12,
+    fontWeight: '400',
+    textAlign: 'left',
+    color: 'black',
+    marginBottom: 16,
+  },
+  authorProcess: {
+    fontFamily: 'Avenir',
+    fontSize: 16,
+    fontWeight: '600',
+    textAlign: 'left',
+    color: 'black',
+    marginBottom: 16,
+  },
+  process: {
+    fontFamily: 'Avenir',
+    fontSize: 12,
+    fontWeight: '400',
+    textAlign: 'left',
+    color: 'black',
+    marginBottom: 16,
+  },
+  backToTopButtonText: {
+    fontFamily: 'Avenir',
+    fontSize: 12,
+    fontWeight: '800',
+    textAlign: 'left',
+    color: 'black',
+  },
+});
+
+export default styles;
